fix(schema): add validations for the incorrect requests

+++ src/services/tournament.services.js
@@ -1,17 +1,25 @@
 import { pool } from '../db/index.js';
 import { uploadImage } from '../lib/cloudinary.js';
-import fs from 'fs-extra'
+import fs from 'fs-extra';
 
-export const newTournamentService = async (tournamentData, tournamentImage, creatorId) => {
+export const newTournamentService = async (
+    tournamentData,
+    tournamentImage,
+    creatorId,
+) => {
     try {
-        const { name, players_count, teams_count, type_tournament } = tournamentData;
-
+        const { name, players_count, teams_count, type_tournament } =
+            tournamentData;
+        if (!creatorId)
+            throw new Error('El ID del creador no puede estar vacío.');
+        if (!name || !players_count || !teams_count || !type_tournament)
+            throw new Error('Todos los campos son requeridos.');
         let logo = null;
         if (tournamentImage) {
             const result = await uploadImage(tournamentImage.tempFilePath);
             logo = result.secure_url;
 
-            await fs.unlink(tournamentImage.tempFilePath)
+            await fs.unlink(tournamentImage.tempFilePath);
         }
 
         const query = `
@@ -19,7 +27,14 @@ export const newTournamentService = async (tournamentData, tournamentImage, crea
         VALUES (?, ?, ?, ?, ?, ?)
         `;
 
-        const result = await pool.query(query, [logo, name, players_count, teams_count, type_tournament, creatorId]);
+        const result = await pool.query(query, [
+            logo,
+            name,
+            players_count,
+            teams_count,
+            type_tournament,
+            creatorId,
+        ]);
 
         return result;
     } catch (error) {
@@ -30,7 +45,9 @@ export const newTournamentService = async (tournamentData, tournamentImage, crea
 
 export const getTournamentsBycreatorIdService = async (creator_id) => {
     try {
-
+        if (!creator_id) {
+            throw new Error('El ID del creador no puede estar vacío.');
+        }
         const query = `
             SELECT * FROM tournament
             WHERE creator_id = ?
@@ -56,9 +73,11 @@ export const getAllTournamentsService = async () => {
     }
 };
 
-
 export const getTournamentsByUserIdService = async (userId, res) => {
     try {
+        if (!userId) {
+            throw new Error('El ID del usuario no puede estar vacío.');
+        }
         const query = `
             SELECT * FROM tournament
             WHERE creator_id = ?
@@ -74,6 +93,9 @@ export const getTournamentsByUserIdService = async (userId, res) => {
 
 export const closeTournamentService = async (tournamentId) => {
     try {
+        if (!tournamentId) {
+            throw new Error('El ID del torneo no puede estar vacío.');
+        }
         const query = `
             UPDATE tournament
             SET finished = true
@@ -86,7 +108,7 @@ export const closeTournamentService = async (tournamentId) => {
         console.error('Error al cerrar el torneo:', error);
         throw new Error('Error interno del servidor');
     }
-}
+};
 
 export const getTournamentTeamsService = async () => {
     try {
@@ -98,16 +120,20 @@ export const getTournamentTeamsService = async () => {
         console.error('Error al obtener los equipos del torneo:', error);
         throw new Error('Error interno del servidor');
     }
-}
+};
 
 export const indexTeamToTournamentService = async (data, logoImage) => {
     try {
         if (!data.tournament_id) {
-            throw new Error('El ID del torneo proporcionado no puede estar vacío.');
+            throw new Error(
+                'El ID del torneo proporcionado no puede estar vacío.',
+            );
         }
 
         const teamExistsQuery = `SELECT id FROM teams WHERE id = ?`;
-        const [teamExistsResult] = await pool.query(teamExistsQuery, [data.team_id]);
+        const [teamExistsResult] = await pool.query(teamExistsQuery, [
+            data.team_id,
+        ]);
 
         if (teamExistsResult.length === 0) {
             let logo_url = null;
@@ -115,32 +141,49 @@ export const indexTeamToTournamentService = async (data, logoImage) => {
                 const result = await uploadImage(logoImage.tempFilePath);
                 logo_url = result.secure_url;
 
-                await fs.unlink(logoImage.tempFilePath)
+                await fs.unlink(logoImage.tempFilePath);
             }
             const insertTeamQuery = `INSERT INTO teams (creator_id, name, logo_url) VALUES (?, ?, ?)`;
-            const [results] = await pool.query(insertTeamQuery, [data.creator_id, data.name, logo_url]);
+            const [results] = await pool.query(insertTeamQuery, [
+                data.creator_id,
+                data.name,
+                logo_url,
+            ]);
             const teamId = results.insertId;
-            console.log("Id del nuevo team: ", teamId);
-            
-            return { teamId, message: 'Equipo creado y agregado al torneo exitosamente' };
+            console.log('Id del nuevo team: ', teamId);
+
+            return {
+                teamId,
+                message: 'Equipo creado y agregado al torneo exitosamente',
+            };
         }
 
         const tournamentExistsQuery = `SELECT id FROM tournament WHERE id = ?`;
-        const [tournamentExistsResult] = await pool.query(tournamentExistsQuery, [data.tournament_id]);
+        const [tournamentExistsResult] = await pool.query(
+            tournamentExistsQuery,
+            [data.tournament_id],
+        );
 
         if (tournamentExistsResult.length === 0) {
             throw new Error('El ID del torneo proporcionado no existe.');
         }
 
         const query = `INSERT INTO tournament_teams (team_id, tournament_id) VALUES (?, ?)`;
-        const result = await pool.query(query, [data.team_id, data.tournament_id]);
+        const result = await pool.query(query, [
+            data.team_id,
+            data.tournament_id,
+        ]);
 
         return result;
     } catch (error) {
         throw new Error(error.message);
     }
-}
+};
 
-
-// Otras funciones para obtener, actualizar o eliminar torneos según sea necesario...
-//logica para poder cerrar un torneo..."finished:bool"
+export const getInfoTournamentService = async (id) => {
+    const query = `
+            SELECT * FROM game_team_stats
+            WHERE game_id IN (SELECT id FROM games WHERE tournament_id = ?)`;
+    const result = await pool.query(query, [id]);
+    return result;
+};
